refactor: compile mojo's to own namespace
