+++ src/main/java/ch/ivyteam/ivy/maven/MavenDependencyMojo.java
@@ -28,6 +28,7 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
 
+import ch.ivyteam.ivy.maven.compile.AbstractProjectCompileMojo;
 import ch.ivyteam.ivy.maven.engine.MavenProjectBuilderProxy;
 import ch.ivyteam.ivy.maven.util.MavenDependencies;
